Enable desktop-logoff sound
The desktop-logoff sound was now added to Yaru (see
https://github.com/ubuntu/yaru/commit/31baf955998528f187253d8f3adbe895501bfc0d).
Now, we can actually use it and move a step closer to completing our sound collection.

This may cause a delay in the shutdown sequence, cinnamon-session waiting
for the sound to finish before it logs out the user. But this can be patched
and if not, disabled.